% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sml_intertidal_cover.R
\docType{data}
\name{sml_intertidal_cover}
\alias{sml_intertidal_cover}
\title{Cover of subtidal species from the SML Intertidal sampling program}
\format{
A data frame with 724,562 rows and 14 variables:
\describe{
\item{site}{The site name}
\item{subtidal_transect}{The transect common name}
...
}
}
\source{
\url{https://doi.org/10.6073/pasta/b8f52444f86fbefe97c5bb40673e02d5.}
}
\usage{
sml_intertidal_cover
}
\description{
A dataset containing percent cover data from the intertidal of SML
}
\keyword{datasets}
