% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sml_intertidal_species.R
\docType{data}
\name{sml_intertidal_species}
\alias{sml_intertidal_species}
\title{Taxonomic information for species sampled intertidally}
\format{
A data frame with 174 rows and 14 variables:
\describe{
\item{Organism}{price, in US dollars}
\item{type}{Invertebrate, Algae, Fungus, Substrate, Detritus}
...
}
}
\source{
\url{https://doi.org/10.6073/pasta/b8f52444f86fbefe97c5bb40673e02d5.}
}
\usage{
sml_intertidal_species
}
\description{
A dataset containing information on taxonomy and AphiaID of all species
sampled by the SML
}
\keyword{datasets}
